services/health_check/requirements.txt
google-cloud-error-reporting==1.0.*
firebase-admin==5.0.*
flask==1.1.*
gunicorn==20.1.*
google-cloud-secret-manager==2.4.*
google-cloud-error-reporting==1.0.*
firebase-admin==5.0.*